Move FlexStateManager to GUI project
